split the localanswerman into its own file
