+++ ManzaHealth.html
@@ -72,6 +72,7 @@
 
 	<script src="js/eventmanager.js"></script>
 	<script src="js/answerman.js"></script>
+	<script src="js/localanswerman.js"></script>
 	<script src="js/submitmanager.js"></script>
     <script src="js/mortar-base.js"></script>
 	<script src="js/widget-base.js"></script>
