+++ js/build.sh
@@ -64,6 +64,7 @@ BRIX_ARGS=$(cat <<EOF
 	--input=bricworks.js
 	--input=domhelper.js
 	--input=eventmanager.js
+	--input=localanswerman.js
 	--input=mortar-base.js
 	--input=mortar-hilite.js
 	--input=ipc.js
