+++ test/index.html
@@ -34,6 +34,7 @@
 	<script src="../js/fakeactivitydb.js"></script>
     <script src="../js/ipsproxy.js"></script>
 	<script src="../js/answerman.js"></script>
+	<script src="../js/localanswerman.js"></script>
 	<script src="../js/submitmanager.js"></script>
 	<script src="../js/widget-base.js"></script>
     <script src="../js/mortar-base.js"></script>
